pkg/endpoint: move WaitForPolicyRevision to policyrevision.go
Signed-off by: Ian Vernon <ian@cilium.io>
