+++ pkg/endpoint/policyrevision.go
@@ -1,0 +1,41 @@
+// Copyright 2016-2018 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package endpoint
+
+import "context"
+
+// WaitForPolicyRevision returns a channel that is closed when one or more of
+// the following conditions have met:
+//  - the endpoint is disconnected state
+//  - the endpoint's policy revision reaches the wanted revision
+func (e *Endpoint) WaitForPolicyRevision(ctx context.Context, rev uint64) <-chan struct{} {
+	e.Mutex.Lock()
+	defer e.Mutex.Unlock()
+	ch := make(chan struct{})
+	if e.policyRevision >= rev || e.state == StateDisconnected {
+		close(ch)
+		return ch
+	}
+	ps := policySignal{
+		wantedRev: rev,
+		ctx:       ctx,
+		ch:        ch,
+	}
+	if e.policyRevisionSignals == nil {
+		e.policyRevisionSignals = map[policySignal]bool{}
+	}
+	e.policyRevisionSignals[ps] = true
+	return ch
+}
